+++ components/molecules/DashboardScatterChart/dashboard-scatter-chart.tsx
@@ -1,6 +1,9 @@
-import React from "react";
+import React, { useState } from "react";
 import EChartWrapper from "components/atoms/EChartWrapper/echart-wrapper";
-import ComponentHeader from "../ComponentHeader/component-header";
+import ToggleOption from "components/atoms/ToggleOption/toggle-option";
+import Text from "components/atoms/Typography/text";
+import Title from "components/atoms/Typography/title";
+
 
 interface DashboardScatterChartProps {
     title: string;
@@ -8,11 +11,39 @@ interface DashboardScatterChartProps {
 }
 
 const DashboardScatterChart: React.FC<DashboardScatterChartProps> = ({ title, option }) => {
+  const [showMembers, setShowMembers] = useState(false);
+  const [showBots, setShowBots] = useState(false);
+  const handleShowMembers = () => {
+    setShowMembers(!showMembers);
+    // Additional logic for showing members
+  };
+  const handleShowBots = () => {
+    setShowBots(!showBots);
+    // Additional logic for showing bots
+  };
   return (
-    <>
-      <ComponentHeader title={title} />
-      <EChartWrapper option={option} />
-    </>
+    <div>
+      <div className="flex justify-between">
+        <Title level={4}>{title}</Title>
+        <div className="flex flex-col md:flex-row gap-2 pr-10">
+          <ToggleOption handleToggle={handleShowBots} checked={showBots} optionText="Show Bots"></ToggleOption>
+          <ToggleOption
+            handleToggle={handleShowMembers}
+            withIcon={true}
+            optionText="Show Outside Contributors"
+            checked={showMembers}
+          ></ToggleOption>
+        </div>
+      </div>
+      <div className="mt-3">
+        <Text className="-rotate-90  opacity-0 md:opacity-100 -left-6 text-light-slate-10 top-[50%] absolute ">
+          Lines Touched
+        </Text>
+        <div className="ml-5">
+          <EChartWrapper option={option} />
+        </div>
+      </div>
+    </div>
   );
 };