+++ components/organisms/Dashboard/dashboard.tsx
@@ -7,11 +7,48 @@ import { useEffect, useState } from "react";
 
 export const Dashboard = (): JSX.Element => {
   const { meta, isError } = useRepositoriesList();
-  const [ itemCountText, setItemCountText ] = useState("Loading...");
-  
+
+  const [itemCountText, setItemCountText] = useState("Loading...");
+
   const scatterOptions = {
-    xAxis: {},
-    yAxis: {},
+    grid: {
+      left: 40,
+      top: 10,
+      right: 40,
+      bottom: 20
+    },
+    xAxis: {
+      boundaryGap: false,
+      scale: true,
+      minInterval: 7,
+      maxInterval: 7,
+      min: 0,
+      max: 35,
+      axisLabel: {
+        formatter: (value: number, index: number) =>
+          value === 0 ? "Today" : value === 35 ? "35+ days ago" : `${value} days ago`
+      },
+      splitLine: {
+        lineStyle: {
+          type: "dashed"
+        }
+      }
+    },
+    yAxis: {
+      min: 0,
+      max: 10000,
+      splitNumber: 6,
+      boundaryGap: false,
+      axisLabel: {
+        showMinLabel: true,
+        formatter: (value: number) => value >= 1000 ? humanizeNumber(value,null) : value
+      },
+      splitLine: {
+        lineStyle: {
+          type: "dashed"
+        }
+      }
+    },
     series: [
       {
         symbolSize: 20,
@@ -45,7 +82,8 @@ export const Dashboard = (): JSX.Element => {
   };
 
   useEffect(() => {
-    if(meta) setItemCountText(`of ${humanizeNumber(meta.itemCount)}`);
+
+    if(meta) setItemCountText(`of ${humanizeNumber(meta.itemCount, "comma")}`);
     if(isError) setItemCountText("of unknown...");
   }, [ isError, meta ]);
 
@@ -94,9 +132,9 @@ export const Dashboard = (): JSX.Element => {
         />
       </section>
       <section className="flex flex-col lg:flex-row max-w-full gap-4 mb-6">
-        <div className="flex flex-col w-full gap-4">
-          <Card className="w-full p-5">
-            <DashboardScatterChart title="Test Title" option={scatterOptions} />
+        <div className="flex flex-col w-full">
+          <Card className="w-full">
+            <DashboardScatterChart title="Contributor Distribution" option={scatterOptions} />
           </Card>
         </div>
       </section>