+++ components/organisms/ToolList/nav.tsx
@@ -39,9 +39,9 @@ const Nav: React.FC<NavProps> = ({ toolList, selectedTool = "dashboard", selecte
                 {tool.name}
               </span>
               {
-                tool.numOf &&
+                (!!tool.numOf || tool.numOf === 0) &&
                   <div className="ml-2 py-0.5 px-1.5 h-fit bg-slate-200 text-slate-500 border rounded-full text-xs font-semibold">
-                    {humanizeNumber(tool.numOf)}
+                    {humanizeNumber(tool.numOf, null)}
                   </div>
               }
             </div>