+++ stories/organisms/contributors.stories.tsx
@@ -1,7 +1,7 @@
 import Contributors from "components/organisms/Contributors/contributors";
 
 const StoryConfig = {
-  title: "Design System/Organisms/Contributors",
+  title: "Design System/Organisms/Contributors"
 };
 export default StoryConfig;